Add an in-game debug menu
This extends the game with an overlaying menu that displays a list of
information about the currently selected object, allowing real-time
observation of changes in its attributes:
-While holding select, use directional buttons to navigate the menu.
Hold select and press up/down to select the next/previous menu item.
When a menu item is selected, hold select and press left or right to
cycle through choices for that item.
-Currently only the first menu and third menu items actually do
something: with the first item selected, one can cycle through the
object pool to select a new object. With the third item selected, one
can cycle through the sub-menus to view additional information about the
currently selected object.
-This menu is currently incomplete; its functionality will be extended
in the future.